pyproject.toml
[project]
name = "adk-samples"
version = "0.1"
description = "C-SAW Agents Template"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "google-adk>=0.3.0",
    "praw>=7.8.1", #for Reddit Scout
    "yfinance>=0.2.55", #for Stock Agent
    "immutabledict>=4.2.1,<5", #for DS Agent
    "sqlglot>=26.10.1,<27", #for DS Agent
    "db-dtypes>=1.4.2,<2", #for DS Agent
    "regex>=2024.11.6,<2025", #for DS Agent
    "tabulate>=0.9.0,<0.10", #for DS Agent
    "absl-py>=2.2.2", #for deployment
    "google-cloud-aiplatform[adk,agent-engines]>=1.91.0", #for deployment scripts
    "prompt-toolkit>=3.0.51", #for interactive scripts
    "nicegui>=2.16.0",
    "google-api-core>=2.24.2",
]

[tool.ruff.lint]
extend-select = ["E", "W", "I"]
ignore = ["E501", "W293", "E701", "E702"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
